+++ src/style.css
@@ -61,7 +61,6 @@ button {
 
 .list-class {
   display: flex;
-  flex-wrap: wrap;
   justify-content: space-between;
   padding: 5px;
 }
@@ -120,4 +119,9 @@ button {
   width: 37px;
   height: 20px;
   margin-top: 2%;
+}
+
+.cancel-text {
+  text-decoration: line-through;
+  color: aqua;
 }
